pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Andrew Lucas <andrewlucasgs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
alembic = "^1.6.5"
fastapi = "^0.65.1"
uvicorn = "^0.14.0"
python-jose = "^3.2.0"
passlib = "^1.7.4"
pydantic = {extras = ["email", "dotenv"], version = "^1.8.2"}
emails = "^0.6"
python-multipart = "^0.0.5"
psycopg2 = "^2.8.6"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
